Split input and output conversions to their own modules
