+++ RecoTracker/MkFit/python/customizeInitialStepOnly.py
@@ -13,8 +13,9 @@ def customizeInitialStepOnly(process):
     process.initialStepHitQuadruplets.SeedComparitorPSet.clusterShapeCacheSrc = "siPixelClusterShapeCachePreSplitting"
     process.initialStepSeeds.SeedComparitorPSet.ClusterShapeCacheSrc = "siPixelClusterShapeCachePreSplitting"
     if hasattr(process.initialStepTrackCandidates, "measurementTrackerEvent"):
+        # MkFit case
         process.initialStepTrackCandidates.measurementTrackerEvent = 'MeasurementTrackerEventPreSplitting'
-        process.initialStepTrackCandidates.pixelRecHits = "siPixelRecHitsPreSplitting"
+        process.initialStepTrackCandidatesMkFitInput.pixelRecHits = "siPixelRecHitsPreSplitting"
     else:
         process.initialStepTrackCandidates.MeasurementTrackerEvent = 'MeasurementTrackerEventPreSplitting'
     process.initialStepTracks.MeasurementTrackerEvent = 'MeasurementTrackerEventPreSplitting'
