+++ RecoTracker/MkFit/src/MkFitEventWrapper.cc
@@ -1,0 +1,1 @@
+#include "RecoTracker/MkFit/interface/MkFitIndexLayer.h"
