+++ RecoTracker/MkFit/src/classes.h
@@ -1,0 +1,8 @@
+#ifndef RecoTracker_MkFit_classes_h
+#define RecoTracker_MkFit_classes_h
+
+#include "DataFormats/Common/interface/Wrapper.h"
+#include "RecoTracker/MkFit/interface/MkFitInputWrapper.h"
+#include "RecoTracker/MkFit/interface/MkFitOutputWrapper.h"
+
+#endif
