+++ Validation/RecoTrack/python/plotting/trackingPlots.py
@@ -1451,6 +1451,9 @@ _iterations = [
                          "initialStepClassifier3",
                          "initialStep",
                          "initialStepSelector"],
+              building=["initialStepTrackCandidatesMkFitInput",
+                        "initialStepTrackCandidatesMkFit",
+                        "initialStepTrackCandidates"],
               other=["firstStepPrimaryVerticesUnsorted",
                      "initialStepTrackRefsForJets",
                      "caloTowerForTrk",
